feat: 表结构优化

+++ src/models/sysUser.ts
@@ -31,7 +31,7 @@ export default function (sequelize: Sequelize) {
             field: 'nickname'
         },
         gender: {
-            type: DataTypes.INTEGER(1),
+            type: DataTypes.INTEGER,
             allowNull: true,
             defaultValue: '1',
             primaryKey: false,
@@ -76,7 +76,7 @@ export default function (sequelize: Sequelize) {
             field: 'mobile'
         },
         status: {
-            type: DataTypes.INTEGER(1),
+            type: DataTypes.INTEGER,
             allowNull: true,
             defaultValue: '1',
             primaryKey: false,
